Consolidate all API interfaces into a single cross-compiled module.
The whole structure of the project is becoming quite entangled in between the modules. Besides, there are modules that do not provide much value and it's naming is wrong with its purpose. Specially `console-shared` and `console-jvm`.

This work is meant to consolidate the actual shared classes for the public API into a single module, containing not just the interfaces but also the public protocols and simplify the console project structure. Server and clients (both AJAX-HTTP and TCP) will provide specialised implementations of the same interface. Pure server-side components meant to support the API will be moved to the master module. Protocol classes will be distributed in packages instead of singleton objects (to ease Akka serialization configuration). Public APIs that require the user to be authenticated will get their signatures modified to reflect so, making it impossible to call an authenticated API without having authentication information in the current scope.